IMPALA-10715: test decimal min max filters failed in exhaustive run
This patch enables only the min/max filters in decimal min/max filter
test so that some of the non-qualifying rows can be returned from the
kudu scanners. Previously, the tests allows bloom filters to filter
out rows at the kudu scanner level which prevents non-qualfying rows
to arrive at the hash join node. Such non-qualifying rows are required
by the test.

With the patch, the test passes in the exhaustive mode. The patch also
refactors the above logic for the entire TestMinMaxFilters test so that
each test case in it will only get the min/max filter.

Testing:
 - Unit test
 - Core test

Change-Id: I20da28f780a27c6fdd917116e7c14d46d2a5db0f
Reviewed-on: http://gerrit.cloudera.org:8080/19132
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>