IMPALA-11657: Ignore git-reset failures in build-all-flag-combinations.sh
When building from a tarball, the git-reset command in
build-all-flag-combinations.sh will fail since it's executed not in a
git repository. The purpose of the command is to revert the changes made
by "mvn versions:set". It's ok to skip this step when building in a
Jenkins job, since that's the last build to verify. No following builds
will be impacted.

This patch ignores the failure of git-reset. So we can set up a Jenkins
job to run build-all-flag-combinations.sh from a tarball.

Tests:
 - Verified the script from a tarball locally.

Change-Id: I2079de0b1eb11044d5293546fe6641939d978134
Reviewed-on: http://gerrit.cloudera.org:8080/19135
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>