IMPALA-11661: Added new api in MetastoreServiceHandler for find_next_compact2 method
In the MetastoreServiceHandler class, the 'find_next_compact2'
API implementation is missing which is causing test failures.
This needs to be fixed by adding implementation of the
'find_next_compact2' api in the MetastoreServiceHandler class

Testing:
Added a test 'test_compaction_apis()' to verify that the
'find_next_compact2' api in HMS is reachable from impala.
The same test can be used in the future to test newly added
apis in HMS.

Change-Id: I9f1663c16d2649c9c455e6dffde02894819b2761
Reviewed-on: http://gerrit.cloudera.org:8080/19140
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>