IMPALA-11524: Bump up the CDP GBN to 33375775 and remove workarounds
This patch bumps up the GBN to 33375775, which contains the
fix for HADOOP-18410 as well as HADOOP-18456, which should help
with the stability of S3 tests. This removes the workaround
for HADOOP-18410 introduced in IMPALA-11514.

This also picks up newer versions of Ozone, Iceberg, Hive, etc.
After HIVE-26071, hive-standalone-metastore starts relying on
jetty-servlet, which in turn requires jetty-security. Since Impala bans
all jetty-related dependencies unless otherwise specified, this patch
adds jetty-servlet, jetty-security, and jetty-util-ajax as allowed
dependencies in order to compile Impala with Hive.

Hive Metastore introduced several new APIs in their interface, so
this adds implementations for these (e.g. HIVE-25303, HIVE-22782).
For example, HIVE-26149 introduced a new HMS API drop_database_req,
so we need to implement this API in CatalogServerHandler as well.
This patch also fixes a bug that the third parameter in drop_database
was treated as "ignoreUnknownDb", however it should be "cascade".

The number of files produced on systems with 32GB of memory changed.
Until we track down the specific cause, this tunes the YARN memory
slightly higher to avoid changes in the number of files produced
for systems with 32GB of memory.

The content in this patch comes from several developers: Yu-Wen Lai,
Kishen Das, and Fang-Yu Rao.

Testing:
 - Ran core jobs
 - Ran CatalogHmsSyncToLatestEventIdTest tests
 - Hand tested on a machine with 32GB of memory

Change-Id: Iea7e1785f5c93f61179cc336968c3a86c53e9ed1
Reviewed-on: http://gerrit.cloudera.org:8080/19149
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Csaba Ringhofer <csringhofer@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>