IMPALA-11673: Exclude spring-jcl from the Java build
spring-core has a dependency on spring-jcl, which is
a jar that implements the same classes as commons-logging.
Having both on the classpath can lead to conflicts and
errors. Since commons-logging is already on the classpath
via other dependencies, this excludes spring-jcl and adds
it to the list of banned dependencies.

Testing:
 - Ran core job
 - Verified spring-jcl is not present

Change-Id: Ifeb741788662ec1b46303154b40109a4eef67005
Reviewed-on: http://gerrit.cloudera.org:8080/19154
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>