IMPALA-11626: Handle COMMIT_COMPACTION_EVENT from HMS
Since HIVE-24329 HMS emits an event when a compaction is committed,
but Impala ignores it. Handling it would allow automatic refreshing
of file metadata after commit compactions.

Testing: Added an end-to-end test that tests the processing of
commit compaction event that was triggered in HMS. Also added an
edge case where event processor would handle the condition of the
partition being missed/deleted.

Change-Id: I464faedb4e3bbcd417bab2e3cb0d57e339d42605
Reviewed-on: http://gerrit.cloudera.org:8080/19155
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>