IMPALA-11695: Reduce clang tidy warning output size
The Clang Tidy build enables all warnings via -Wall
and -Weverything. This produces enormous output.
Looking at a recent failed Clang Tidy build, there
are ~4.5 million warnings generated. Of these,
about 4 million are from C++98 compatibility warnings.
A further 250 thousand are from padding warnings.
Since these are not particularly interesting, this
disables both of those to reduce the output size.

Disabling these warnings allowed Clang Tidy to find
some issues in DataSketches that it was previously
missing. Perhaps there is some limit on the number
or size of warnings that it was processing. This
modifies the DataSketches code to fix those (which
are all minor issues with const correctness).

Testing:
 - Built with clang tidy locally

Change-Id: I28c6ed1e7a4f525d81a9c48e90d051b374d44941
Reviewed-on: http://gerrit.cloudera.org:8080/19182
Reviewed-by: Wenzhe Zhou <wzhou@cloudera.com>
Reviewed-by: Michael Smith <michael.smith@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>