IMPALA-11717: Use rapidjson for printing collections
We have been using rapidjson to print structs but didn't use it to print
collections (arrays and maps).

This change introduces the usage of rapidjson to print collections for
both the HS2 and the Beeswax protocol.

The old code handling the printing of collections in raw-value.{h,cc} is
removed.

Testing:
 - Ran existing EE tests
 - Added EE tests with non-string and NULL map keys in
   nested-map-in-select-list.test and map_null_keys.test.

Change-Id: I08a2d596a498fbbaf1419b18284846b992f49165
Reviewed-on: http://gerrit.cloudera.org:8080/19309
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Daniel Becker <daniel.becker@cloudera.com>